FIX: Visual C++ warnings
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@3976 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/apps/mrbias_cuda.cxx
@@ -177,7 +177,7 @@ doMain( const int argc, const char *argv[] )
     {
     if ( ThresholdAuto )
       {
-      ThresholdForegroundMin = cmtk::TypedArrayNoiseEstimatorNaiveGaussian( *(inputImage->GetData()) ).GetNoiseThreshold();
+      ThresholdForegroundMin = static_cast<float>( cmtk::TypedArrayNoiseEstimatorNaiveGaussian( *(inputImage->GetData()) ).GetNoiseThreshold() );
       ThresholdForegroundFlag = true;
 
       cmtk::DebugOutput( 1 ) << "INFO: estimated foreground threshold from noise level as " << ThresholdForegroundMin << "\n";
